Small changes in krb: 1 signed/unsigned, 1 disabling stderr, and changed LIB name
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@5078 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/chk_trans.c
@@ -66,7 +66,7 @@ krb5_data      *realm2;
         if (*(nextp-1) == '.')  strcat(next, prev);
         retval = KRB5KRB_AP_ERR_ILL_CR_TKT;
         for (j = 0; tgs_list[j]; j++) {
-          if (strlen(next) == krb5_princ_realm(context, tgs_list[j])->length &&
+          if (strlen(next) == (size_t) krb5_princ_realm(context, tgs_list[j])->length &&
               !memcmp(next, krb5_princ_realm(context, tgs_list[j])->data,
                       strlen(next))) {
             retval = 0;
